+++ tests/DatabaseRegistryTest.php
@@ -25,7 +25,7 @@ class MockMapRegistry extends DatabaseRegistry
     protected string $tableName = 'MockMap';
     protected string $primaryKey = 'id';
 
-    public function createTable(Database $database): void
+    public function createTable(Database $database = null): void
     {
         $helper = $database->getHelper();
         $database->execute(